Pull out test utils
