+++ impl/src/main/java/org/wildfly/extras/a2a/server/apps/jakarta/TestingHook.java
@@ -1,0 +1,4 @@
+package org.wildfly.extras.a2a.server.apps.jakarta;
+
+public class TestingHook {
+}
