+++ test-utils/src/main/java/org/wildfly/extras/a2a/server/jakarta/test/utils/RestApplication.java
@@ -1,4 +1,4 @@
-package org.wildfly.extras.a2a.server.apps.jakarta;
+package org.wildfly.extras.a2a.server.jakarta.test.utils;
 
 import jakarta.ws.rs.ApplicationPath;
 import jakarta.ws.rs.core.Application;
