+++ wildfly-jar/pom.xml
@@ -62,6 +62,16 @@
             </exclusions>
         </dependency>
         <dependency>
+            <groupId>io.smallrye.reactive</groupId>
+            <artifactId>mutiny-zero</artifactId>
+            <exclusions>
+                <exclusion>
+                    <groupId>*</groupId>
+                    <artifactId>*</artifactId>
+                </exclusion>
+            </exclusions>
+        </dependency>
+        <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>a2a-java-sdk-server-jakarta</artifactId>
             <version>${project.version}</version>
